Move io into io subdir
