+++ configure.ac
@@ -287,18 +287,18 @@ AC_CONFIG_LINKS([include/souffle/ExplainProvenance.h:src/include/souffle/Explain
 AC_CONFIG_LINKS([include/souffle/ExplainProvenanceImpl.h:src/include/souffle/ExplainProvenanceImpl.h])
 AC_CONFIG_LINKS([include/souffle/ExplainTree.h:src/include/souffle/ExplainTree.h])
 AC_CONFIG_LINKS([include/souffle/EquivalenceRelation.h:src/include/souffle/EquivalenceRelation.h])
-AC_CONFIG_LINKS([include/souffle/IOSystem.h:src/include/souffle/IOSystem.h])
+AC_CONFIG_LINKS([include/souffle/io/IOSystem.h:src/include/souffle/io/IOSystem.h])
 AC_CONFIG_LINKS([include/souffle/LambdaBTree.h:src/include/souffle/LambdaBTree.h])
 AC_CONFIG_LINKS([include/souffle/Logger.h:src/include/souffle/Logger.h])
 AC_CONFIG_LINKS([include/souffle/PiggyList.h:src/include/souffle/PiggyList.h])
 AC_CONFIG_LINKS([include/souffle/ProfileDatabase.h:src/include/souffle/ProfileDatabase.h])
 AC_CONFIG_LINKS([include/souffle/ProfileEvent.h:src/include/souffle/ProfileEvent.h])
 AC_CONFIG_LINKS([include/souffle/RamTypes.h:src/include/souffle/RamTypes.h])
-AC_CONFIG_LINKS([include/souffle/ReadStream.h:src/include/souffle/ReadStream.h])
-AC_CONFIG_LINKS([include/souffle/ReadStreamCSV.h:src/include/souffle/ReadStreamCSV.h])
-AC_CONFIG_LINKS([include/souffle/ReadStreamSQLite.h:src/include/souffle/ReadStreamSQLite.h])
-AC_CONFIG_LINKS([include/souffle/ReadStreamJSON.h:src/include/souffle/ReadStreamJSON.h])
-AC_CONFIG_LINKS([include/souffle/SerialisationStream.h:src/include/souffle/SerialisationStream.h])
+AC_CONFIG_LINKS([include/souffle/io/ReadStream.h:src/include/souffle/io/ReadStream.h])
+AC_CONFIG_LINKS([include/souffle/io/ReadStreamCSV.h:src/include/souffle/io/ReadStreamCSV.h])
+AC_CONFIG_LINKS([include/souffle/io/ReadStreamSQLite.h:src/include/souffle/io/ReadStreamSQLite.h])
+AC_CONFIG_LINKS([include/souffle/io/ReadStreamJSON.h:src/include/souffle/io/ReadStreamJSON.h])
+AC_CONFIG_LINKS([include/souffle/io/SerialisationStream.h:src/include/souffle/io/SerialisationStream.h])
 AC_CONFIG_LINKS([include/souffle/SignalHandler.h:src/include/souffle/SignalHandler.h])
 AC_CONFIG_LINKS([include/souffle/SouffleInterface.h:src/include/souffle/SouffleInterface.h])
 AC_CONFIG_LINKS([include/souffle/SymbolTable.h:src/include/souffle/SymbolTable.h])
@@ -316,11 +316,11 @@ AC_CONFIG_LINKS([include/souffle/utility/StreamUtil.h:src/include/souffle/utilit
 AC_CONFIG_LINKS([include/souffle/utility/StringUtil.h:src/include/souffle/utility/StringUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/ParallelUtil.h:src/include/souffle/utility/ParallelUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/tinyformat.h:src/include/souffle/utility/tinyformat.h])
-AC_CONFIG_LINKS([include/souffle/WriteStream.h:src/include/souffle/WriteStream.h])
-AC_CONFIG_LINKS([include/souffle/WriteStreamCSV.h:src/include/souffle/WriteStreamCSV.h])
-AC_CONFIG_LINKS([include/souffle/WriteStreamSQLite.h:src/include/souffle/WriteStreamSQLite.h])
-AC_CONFIG_LINKS([include/souffle/WriteStreamJSON.h:src/include/souffle/WriteStreamJSON.h])
-AC_CONFIG_LINKS([include/souffle/gzfstream.h:src/include/souffle/gzfstream.h])
+AC_CONFIG_LINKS([include/souffle/io/WriteStream.h:src/include/souffle/io/WriteStream.h])
+AC_CONFIG_LINKS([include/souffle/io/WriteStreamCSV.h:src/include/souffle/io/WriteStreamCSV.h])
+AC_CONFIG_LINKS([include/souffle/io/WriteStreamSQLite.h:src/include/souffle/io/WriteStreamSQLite.h])
+AC_CONFIG_LINKS([include/souffle/io/WriteStreamJSON.h:src/include/souffle/io/WriteStreamJSON.h])
+AC_CONFIG_LINKS([include/souffle/io/gzfstream.h:src/include/souffle/io/gzfstream.h])
 AC_CONFIG_LINKS([include/souffle/json11.h:src/include/souffle/json11.h])
 AC_CONFIG_LINKS([include/souffle/profile/Cell.h:src/include/souffle/profile/Cell.h])
 AC_CONFIG_LINKS([include/souffle/profile/CellInterface.h:src/include/souffle/profile/CellInterface.h])
