+++ src/include/souffle/CompiledSouffle.h
@@ -19,14 +19,14 @@
 #include "souffle/Brie.h"
 #include "souffle/CompiledTuple.h"
 #include "souffle/EquivalenceRelation.h"
-#include "souffle/IOSystem.h"
 #include "souffle/RamTypes.h"
 #include "souffle/RecordTable.h"
 #include "souffle/SignalHandler.h"
 #include "souffle/SouffleInterface.h"
 #include "souffle/SymbolTable.h"
 #include "souffle/Table.h"
-#include "souffle/WriteStream.h"
+#include "souffle/io/IOSystem.h"
+#include "souffle/io/WriteStream.h"
 #include "souffle/utility/CacheUtil.h"
 #include "souffle/utility/ContainerUtil.h"
 #include "souffle/utility/EvaluatorUtil.h"
