+++ src/include/souffle/io/ReadStreamJSON.h
@@ -15,8 +15,8 @@
 #pragma once
 
 #include "souffle/RamTypes.h"
-#include "souffle/ReadStream.h"
 #include "souffle/SymbolTable.h"
+#include "souffle/io/ReadStream.h"
 #include "souffle/utility/ContainerUtil.h"
 #include "souffle/utility/FileUtil.h"
 #include "souffle/utility/StringUtil.h"
