+++ src/include/souffle/io/SerialisationStream.h
