+++ src/interpreter/InterpreterEngine.cpp
@@ -77,15 +77,15 @@
 #include "ram/UserDefinedOperator.h"
 #include "ram/Visitor.h"
 #include "souffle/BinaryConstraintOps.h"
-#include "souffle/IOSystem.h"
 #include "souffle/Logger.h"
 #include "souffle/ProfileEvent.h"
 #include "souffle/RamTypes.h"
-#include "souffle/ReadStream.h"
 #include "souffle/RecordTable.h"
 #include "souffle/SignalHandler.h"
 #include "souffle/SymbolTable.h"
-#include "souffle/WriteStream.h"
+#include "souffle/io/IOSystem.h"
+#include "souffle/io/ReadStream.h"
+#include "souffle/io/WriteStream.h"
 #include "souffle/utility/EvaluatorUtil.h"
 #include "souffle/utility/MiscUtil.h"
 #include "souffle/utility/ParallelUtil.h"
